move and update test file
